requirements.txt
fastapi == 0.129.0
uvicorn == 0.34.0
python-multipart == 0.0.22
tensorflow-cpu == 2.20.0
numpy == 1.26.4
pydantic == 2.10.4
pandas == 2.2.3
matplotlib == 3.9.4
pillow >= 12.1.1
networkx == 3.2.1
scipy==1.15.1